// src/components/name-card/types.ts
export interface Person {
  id: number;
  name: string;
  chucVu: string;
  chucDanh: string;
}

export interface CardConfig {
  zoneTop: number;
  zoneBottom: number;
  gapRoleName: number;
  gapNamePos: number;
  fontSizeRole: number;
  fontSizeName: number;
  fontSizePos: number;
  textColor: string;
  fontFamilyRole: string;
  fontFamilyName: string;
  fontFamilyPos: string;
}